+++ .gitignore
@@ -12,6 +12,9 @@ Icon
 # Thumbnails
 ._*
 
+# Turbo
+.turbo
+
 # Files that might appear in the root of a volume
 .DocumentRevisions-V100
 .fseventsd
