+++ .prettierrc
@@ -2,6 +2,7 @@
     "trailingComma": "none",
     "singleQuote": true,
     "printWidth": 80,
-    "tabWidth": 2
+    "tabWidth": 2,
+    "endOfLine": "auto"
 }
