+++ apps/docs/.gitignore
