+++ apps/docs/public/favicon.ico
