+++ apps/docs/public/manifest.json
