+++ apps/docs/tsconfig.json
