--- lerna.json
@@ -1,8 +1,0 @@
-{
-  "packages": [
-    "packages/*"
-  ],
-  "version": "1.2.17-alpha.0",
-  "npmClient": "yarn",
-  "useWorkspaces": true
-}
