+++ packages/components/lib/cjs/components/Container.d.ts
@@ -1,7 +1,8 @@
 import { colorsTokens } from '@irubingu/react-8bit-tokens';
 import { CSS } from '@stitches/react';
-import type { FC } from 'react';
+import type { FC, ReactNode } from 'react';
 declare type React8bitUiContainerType = FC<{
+    children?: ReactNode;
     centered?: boolean;
     borderColor?: keyof typeof colorsTokens;
     bgColor?: CSS['backgroundColor'];
