+++ packages/components/lib/cjs/index.js
@@ -1,2 +1,2 @@
-"use strict";Object.defineProperty(exports,"__esModule",{value:!0});var e=require("react/jsx-runtime"),r=require("@stitches/react"),t=require("polished"),o=function(){return o=Object.assign||function(e){for(var r,t=1,o=arguments.length;t<o;t++)for(var n in r=arguments[t])Object.prototype.hasOwnProperty.call(r,n)&&(e[n]=r[n]);return e},o.apply(this,arguments)};function n(e,r){var t={};for(var o in e)Object.prototype.hasOwnProperty.call(e,o)&&r.indexOf(o)<0&&(t[o]=e[o]);if(null!=e&&"function"==typeof Object.getOwnPropertySymbols){var n=0;for(o=Object.getOwnPropertySymbols(e);n<o.length;n++)r.indexOf(o[n])<0&&Object.prototype.propertyIsEnumerable.call(e,o[n])&&(t[o[n]]=e[o[n]])}return t}var a,i={primary:"rgb(32, 156, 238)",success:"rgb(146, 204, 65)",warning:"rgb(247, 213, 29)",error:"rgb(231, 110, 85)"},c=(a=r.createStitches({theme:{colors:o({},i)},media:{bp1:"(min-width: 480px)",pixelRatio:"(-webkit-min-device-pixel-ratio: 0) and (min-resolution: 0.001dpcm)"},prefix:"react8bit",utils:{marginX:function(e){return{marginLeft:e,marginRight:e}}}})).styled;a.css;var l=a.globalCss;a.keyframes,a.getCssText,a.theme;var s=a.createTheme;function d(e){return Object.keys(e)}a.config;var f={primary:{text:"#ffffff",hover:"#108de0",shade:"#006bb3"},success:{text:"#ffffff",hover:"#76c442",shade:"#4aa52e"},warning:{text:"#212529",hover:"#f2c409",shade:"#e59400"},error:{text:"#fff",hover:"#ce372b",shade:"#8c2022"}},b="#212529",h="#e7e7e7",m="#adafbc",g=c("button",{backgroundColor:"#ffffff",fontSize:"1rem",borderStyle:"solid",borderWidth:4,borderImageSlice:2,borderImageWidth:2,borderImageRepeat:"stretch",borderImageSource:'url(\'data:image/svg+xml;utf8,<?xml version="1.0" encoding="UTF-8" ?><svg version="1.1" width="5" height="5" xmlns="http://www.w3.org/2000/svg"><path d="M2 1 h1 v1 h-1 z M1 2 h1 v1 h-1 z M3 2 h1 v1 h-1 z M2 3 h1 v1 h-1 z" fill="rgb(33,37,41)" /></svg>\')',borderImageOutset:2,position:"relative",display:"inline-block",padding:"6px 8px",margin:4,textAlign:"center",verticalAlign:"middle",cursor:"url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAACAAAAAgCAYAAABzenr0AAAAzElEQVRYR+2X0Q6AIAhF5f8/2jYXZkwEjNSVvVUjDpcrGgT7FUkI2D9xRfQETwNIiWO85wfINfQUEyxBG2ArsLwC0jioGt5zFcwF4OYDPi/mBYKm4t0U8ATgRm3ThFoAqkhNgWkA0jJLvaOVSs7j3qMnSgXWBMiWPXe94QqMBMBc1VZIvaTu5u5pQewq0EqNZvIEMCmxAawK0DNkay9QmfFNAJUXfgGgUkLaE7j/h8fnASkxHTz0DGIBMCnBeeM7AArpUd3mz2x3C7wADglA8BcWMZhZAAAAAElFTkSuQmCC) 14 0,pointer",userSelect:"none","&[href]":{textDecoration:"none"},"&::after":{position:"absolute",top:t.rem(-4),right:t.rem(-4),bottom:t.rem(-4),left:t.rem(-4),content:"",boxShadow:"inset -".concat(t.rem(4)," -").concat(t.rem(4)," ").concat(m)},"&:hover":{backgroundColor:h,"&::after":{boxShadow:"inset -".concat(t.rem(6)," -").concat(t.rem(6)," ").concat(m)}},"&:active":{"&::after":{boxShadow:"inset ".concat(t.rem(4)," ").concat(t.rem(4)," ").concat(m)}},"&[disabled]":{opacity:.5,cursor:"not-allowed",backgroundColor:"#d3d3d3",color:b,"&::after":{boxShadow:"inset -".concat(t.rem(4)," -").concat(t.rem(4)," ").concat(m)}},"@pixelRatio":{borderImageRepeat:"space"},variants:{color:o({},function(){for(var e={},r=0,o=d(i);r<o.length;r++){var n=o[r];e[n]={"&:not([disabled])":{backgroundColor:"$".concat(n),color:f[n].text,"&::after":{boxShadow:"inset -".concat(t.rem(4)," -").concat(t.rem(4)," ").concat(f[n].shade)},"&:hover":{backgroundColor:f[n].hover,"&::after":{boxShadow:"inset -".concat(t.rem(6)," -").concat(t.rem(6)," ").concat(f[n].shade)}},"&:active":{"&::after":{boxShadow:"inset ".concat(t.rem(4)," ").concat(t.rem(4)," ").concat(f[n].shade)}}}}}return e}())}}),p=c("section",{position:"relative",padding:"1.5rem 2rem",borderColor:"#000000",borderStyle:"solid",borderWidth:4,variants:{centered:{true:{textAlign:"center"}},title:{true:{"> .title":{display:"table",padding:"0 .5rem",margin:"-2.3rem 0 1rem",fontSize:"1rem",backgroundColor:"#ffffff"}}},borderColor:o({},function(){for(var e={},r=0,t=d(i);r<t.length;r++){var o=t[r];e[o]={borderColor:"$".concat(o)}}return e}())},compoundVariants:[{centered:!0,title:!0,css:{"> .title":{margin:"-2.3rem auto 1rem"}}}]}),u={html:{fontSize:16,"-webkit-text-size-adjust":"100%","-webkit-tap-highlight-color":"transparent"},body:{margin:0,fontFamily:"'Press Start 2P' sans-serif",lineHeight:1.5,fontSize:16},"*, ::before, ::after":{boxSizing:"border-box",borderStyle:"solid",borderWidth:0},"button, input, optgroup, select, textarea":{"-webkit-appearance":"none",appearance:"none",verticalAlign:"middle",color:"inherit",font:"inherit",background:"transparent",padding:0,margin:0,borderRadius:0,textAlign:"inherit",textTransform:"inherit"},'button, [type="button"], [type="reset"], [type="submit"]':{cursor:"pointer"},fieldset:{margin:0,padding:0,minWidth:0},table:{borderColor:"inherit",borderCollapse:"collapse"},caption:{textAlign:"left"}},v=l(u);exports.Button=function(r){var t=r.children,a=r.color,i=r.href,c=n(r,["children","color","href"]);return e.jsx(g,o({type:i?void 0:"button"},c,{color:a,as:i?"a":"button",href:i&&i},{children:t}),void 0)},exports.Container=function(r){var t=r.children,a=r.centered,i=r.borderColor,c=r.as,l=r.title,s=n(r,["children","centered","borderColor","as","title"]);return e.jsxs(p,o({},s,{borderColor:i,centered:a,title:Boolean(l),as:c},{children:[l&&e.jsx("h3",o({className:"title"},{children:l}),void 0),t]}),void 0)},exports.create8bitTheme=s,exports.globalReact8BitStyles=v,exports.globalReact8BitStylesObj=u;
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0});var e=require("react/jsx-runtime"),r=require("polished"),t=require("@stitches/react"),o=function(){return o=Object.assign||function(e){for(var r,t=1,o=arguments.length;t<o;t++)for(var n in r=arguments[t])Object.prototype.hasOwnProperty.call(r,n)&&(e[n]=r[n]);return e},o.apply(this,arguments)};function n(e,r){var t={};for(var o in e)Object.prototype.hasOwnProperty.call(e,o)&&r.indexOf(o)<0&&(t[o]=e[o]);if(null!=e&&"function"==typeof Object.getOwnPropertySymbols){var n=0;for(o=Object.getOwnPropertySymbols(e);n<o.length;n++)r.indexOf(o[n])<0&&Object.prototype.propertyIsEnumerable.call(e,o[n])&&(t[o[n]]=e[o[n]])}return t}var a,i={primary:"rgb(32, 156, 238)",success:"rgb(146, 204, 65)",warning:"rgb(247, 213, 29)",error:"rgb(231, 110, 85)"},c=(a=t.createStitches({theme:{colors:o({},i)},media:{bp1:"(min-width: 480px)",pixelRatio:"(-webkit-min-device-pixel-ratio: 0) and (min-resolution: 0.001dpcm)"},prefix:"react8bit",utils:{marginX:function(e){return{marginLeft:e,marginRight:e}}}})).styled;a.css;var l=a.globalCss;a.keyframes,a.getCssText,a.theme;var s=a.createTheme;function d(e){return Object.keys(e)}a.config;var f={primary:{text:"#ffffff",hover:"#108de0",shade:"#006bb3"},success:{text:"#ffffff",hover:"#76c442",shade:"#4aa52e"},warning:{text:"#212529",hover:"#f2c409",shade:"#e59400"},error:{text:"#fff",hover:"#ce372b",shade:"#8c2022"}},b="#212529",h="#e7e7e7",m="#adafbc",g=c("button",{backgroundColor:"#ffffff",fontSize:"1rem",borderStyle:"solid",borderWidth:4,borderImageSlice:2,borderImageWidth:2,borderImageRepeat:"stretch",borderImageSource:'url(\'data:image/svg+xml;utf8,<?xml version="1.0" encoding="UTF-8" ?><svg version="1.1" width="5" height="5" xmlns="http://www.w3.org/2000/svg"><path d="M2 1 h1 v1 h-1 z M1 2 h1 v1 h-1 z M3 2 h1 v1 h-1 z M2 3 h1 v1 h-1 z" fill="rgb(33,37,41)" /></svg>\')',borderImageOutset:2,position:"relative",display:"inline-block",padding:"6px 8px",margin:4,textAlign:"center",verticalAlign:"middle",cursor:"url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAACAAAAAgCAYAAABzenr0AAAAzElEQVRYR+2X0Q6AIAhF5f8/2jYXZkwEjNSVvVUjDpcrGgT7FUkI2D9xRfQETwNIiWO85wfINfQUEyxBG2ArsLwC0jioGt5zFcwF4OYDPi/mBYKm4t0U8ATgRm3ThFoAqkhNgWkA0jJLvaOVSs7j3qMnSgXWBMiWPXe94QqMBMBc1VZIvaTu5u5pQewq0EqNZvIEMCmxAawK0DNkay9QmfFNAJUXfgGgUkLaE7j/h8fnASkxHTz0DGIBMCnBeeM7AArpUd3mz2x3C7wADglA8BcWMZhZAAAAAElFTkSuQmCC) 14 0,pointer",userSelect:"none","&[href]":{textDecoration:"none"},"&::after":{position:"absolute",top:r.rem(-4),right:r.rem(-4),bottom:r.rem(-4),left:r.rem(-4),content:"",boxShadow:"inset -".concat(r.rem(4)," -").concat(r.rem(4)," ").concat(m)},"&:hover":{backgroundColor:h,"&::after":{boxShadow:"inset -".concat(r.rem(6)," -").concat(r.rem(6)," ").concat(m)}},"&:active":{"&::after":{boxShadow:"inset ".concat(r.rem(4)," ").concat(r.rem(4)," ").concat(m)}},"&[disabled]":{opacity:.5,cursor:"not-allowed",backgroundColor:"#d3d3d3",color:b,"&::after":{boxShadow:"inset -".concat(r.rem(4)," -").concat(r.rem(4)," ").concat(m)}},"@pixelRatio":{borderImageRepeat:"space"},variants:{color:o({},function(){for(var e={},t=0,o=d(i);t<o.length;t++){var n=o[t];e[n]={"&:not([disabled])":{backgroundColor:"$".concat(n),color:f[n].text,"&::after":{boxShadow:"inset -".concat(r.rem(4)," -").concat(r.rem(4)," ").concat(f[n].shade)},"&:hover":{backgroundColor:f[n].hover,"&::after":{boxShadow:"inset -".concat(r.rem(6)," -").concat(r.rem(6)," ").concat(f[n].shade)}},"&:active":{"&::after":{boxShadow:"inset ".concat(r.rem(4)," ").concat(r.rem(4)," ").concat(f[n].shade)}}}}}return e}())}}),p=c("section",{position:"relative",padding:"1.5rem 2rem",borderColor:"#000000",borderStyle:"solid",borderWidth:4,variants:{centered:{true:{textAlign:"center"}},title:{true:{"> .title":{display:"table",padding:"0 .5rem",margin:"-2.3rem 0 1rem",fontSize:"1rem",backgroundColor:"#ffffff"}}},borderColor:o({},function(){for(var e={},r=0,t=d(i);r<t.length;r++){var o=t[r];e[o]={borderColor:"$".concat(o)}}return e}())},compoundVariants:[{centered:!0,title:!0,css:{"> .title":{margin:"-2.3rem auto 1rem"}}}]}),u={html:{fontSize:16,"-webkit-text-size-adjust":"100%","-webkit-tap-highlight-color":"transparent"},body:{margin:0,fontFamily:"'Press Start 2P' sans-serif",lineHeight:1.5,fontSize:16},"*, ::before, ::after":{boxSizing:"border-box",borderStyle:"solid",borderWidth:0},"button, input, optgroup, select, textarea":{"-webkit-appearance":"none",appearance:"none",verticalAlign:"middle",color:"inherit",font:"inherit",background:"transparent",padding:0,margin:0,borderRadius:0,textAlign:"inherit",textTransform:"inherit"},'button, [type="button"], [type="reset"], [type="submit"]':{cursor:"pointer"},fieldset:{margin:0,padding:0,minWidth:0},table:{borderColor:"inherit",borderCollapse:"collapse"},caption:{textAlign:"left"}},x=l(u);exports.Button=function(r){var t=r.children,a=r.color,i=r.href,c=n(r,["children","color","href"]);return e.jsx(g,o({type:i?void 0:"button"},c,{color:a,as:i?"a":"button",href:i},{children:t}))},exports.Container=function(r){var t=r.children,a=r.centered,i=r.borderColor,c=r.as,l=r.title,s=n(r,["children","centered","borderColor","as","title"]);return e.jsxs(p,o({},s,{borderColor:i,centered:a,title:Boolean(l),as:c},{children:[l&&e.jsx("h3",o({className:"title"},{children:l})),t]}))},exports.create8bitTheme=s,exports.globalReact8BitStyles=x,exports.globalReact8BitStylesObj=u;
 //# sourceMappingURL=index.js.map
