+++ packages/components/lib/esm/index.js
@@ -1,16 +1,2 @@
-import{jsx as e,jsxs as t}from"react/jsx-runtime";import{createStitches as r}from"@stitches/react";import{rem as o}from"polished";
-/*! *****************************************************************************
-Copyright (c) Microsoft Corporation.
-
-Permission to use, copy, modify, and/or distribute this software for any
-purpose with or without fee is hereby granted.
-
-THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES WITH
-REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF MERCHANTABILITY
-AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY SPECIAL, DIRECT,
-INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES WHATSOEVER RESULTING FROM
-LOSS OF USE, DATA OR PROFITS, WHETHER IN AN ACTION OF CONTRACT, NEGLIGENCE OR
-OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR
-PERFORMANCE OF THIS SOFTWARE.
-***************************************************************************** */var n=function(){return n=Object.assign||function(e){for(var t,r=1,o=arguments.length;r<o;r++)for(var n in t=arguments[r])Object.prototype.hasOwnProperty.call(t,n)&&(e[n]=t[n]);return e},n.apply(this,arguments)};function a(e,t){var r={};for(var o in e)Object.prototype.hasOwnProperty.call(e,o)&&t.indexOf(o)<0&&(r[o]=e[o]);if(null!=e&&"function"==typeof Object.getOwnPropertySymbols){var n=0;for(o=Object.getOwnPropertySymbols(e);n<o.length;n++)t.indexOf(o[n])<0&&Object.prototype.propertyIsEnumerable.call(e,o[n])&&(r[o[n]]=e[o[n]])}return r}var i,c={primary:"rgb(32, 156, 238)",success:"rgb(146, 204, 65)",warning:"rgb(247, 213, 29)",error:"rgb(231, 110, 85)"},l=(i=r({theme:{colors:n({},c)},media:{bp1:"(min-width: 480px)",pixelRatio:"(-webkit-min-device-pixel-ratio: 0) and (min-resolution: 0.001dpcm)"},prefix:"react8bit",utils:{marginX:function(e){return{marginLeft:e,marginRight:e}}}})).styled;i.css;var d=i.globalCss;i.keyframes,i.getCssText,i.theme;var s=i.createTheme;function f(e){return Object.keys(e)}i.config;var h={primary:{text:"#ffffff",hover:"#108de0",shade:"#006bb3"},success:{text:"#ffffff",hover:"#76c442",shade:"#4aa52e"},warning:{text:"#212529",hover:"#f2c409",shade:"#e59400"},error:{text:"#fff",hover:"#ce372b",shade:"#8c2022"}},b="#212529",g="#e7e7e7",p="#adafbc",m=l("button",{backgroundColor:"#ffffff",fontSize:"1rem",borderStyle:"solid",borderWidth:4,borderImageSlice:2,borderImageWidth:2,borderImageRepeat:"stretch",borderImageSource:'url(\'data:image/svg+xml;utf8,<?xml version="1.0" encoding="UTF-8" ?><svg version="1.1" width="5" height="5" xmlns="http://www.w3.org/2000/svg"><path d="M2 1 h1 v1 h-1 z M1 2 h1 v1 h-1 z M3 2 h1 v1 h-1 z M2 3 h1 v1 h-1 z" fill="rgb(33,37,41)" /></svg>\')',borderImageOutset:2,position:"relative",display:"inline-block",padding:"6px 8px",margin:4,textAlign:"center",verticalAlign:"middle",cursor:"url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAACAAAAAgCAYAAABzenr0AAAAzElEQVRYR+2X0Q6AIAhF5f8/2jYXZkwEjNSVvVUjDpcrGgT7FUkI2D9xRfQETwNIiWO85wfINfQUEyxBG2ArsLwC0jioGt5zFcwF4OYDPi/mBYKm4t0U8ATgRm3ThFoAqkhNgWkA0jJLvaOVSs7j3qMnSgXWBMiWPXe94QqMBMBc1VZIvaTu5u5pQewq0EqNZvIEMCmxAawK0DNkay9QmfFNAJUXfgGgUkLaE7j/h8fnASkxHTz0DGIBMCnBeeM7AArpUd3mz2x3C7wADglA8BcWMZhZAAAAAElFTkSuQmCC) 14 0,pointer",userSelect:"none","&[href]":{textDecoration:"none"},"&::after":{position:"absolute",top:o(-4),right:o(-4),bottom:o(-4),left:o(-4),content:"",boxShadow:"inset -".concat(o(4)," -").concat(o(4)," ").concat(p)},"&:hover":{backgroundColor:g,"&::after":{boxShadow:"inset -".concat(o(6)," -").concat(o(6)," ").concat(p)}},"&:active":{"&::after":{boxShadow:"inset ".concat(o(4)," ").concat(o(4)," ").concat(p)}},"&[disabled]":{opacity:.5,cursor:"not-allowed",backgroundColor:"#d3d3d3",color:b,"&::after":{boxShadow:"inset -".concat(o(4)," -").concat(o(4)," ").concat(p)}},"@pixelRatio":{borderImageRepeat:"space"},variants:{color:n({},function(){for(var e={},t=0,r=f(c);t<r.length;t++){var n=r[t];e[n]={"&:not([disabled])":{backgroundColor:"$".concat(n),color:h[n].text,"&::after":{boxShadow:"inset -".concat(o(4)," -").concat(o(4)," ").concat(h[n].shade)},"&:hover":{backgroundColor:h[n].hover,"&::after":{boxShadow:"inset -".concat(o(6)," -").concat(o(6)," ").concat(h[n].shade)}},"&:active":{"&::after":{boxShadow:"inset ".concat(o(4)," ").concat(o(4)," ").concat(h[n].shade)}}}}}return e}())}}),u=function(t){var r=t.children,o=t.color,i=t.href,c=a(t,["children","color","href"]);return e(m,n({type:i?void 0:"button"},c,{color:o,as:i?"a":"button",href:i&&i},{children:r}),void 0)},v=l("section",{position:"relative",padding:"1.5rem 2rem",borderColor:"#000000",borderStyle:"solid",borderWidth:4,variants:{centered:{true:{textAlign:"center"}},title:{true:{"> .title":{display:"table",padding:"0 .5rem",margin:"-2.3rem 0 1rem",fontSize:"1rem",backgroundColor:"#ffffff"}}},borderColor:n({},function(){for(var e={},t=0,r=f(c);t<r.length;t++){var o=r[t];e[o]={borderColor:"$".concat(o)}}return e}())},compoundVariants:[{centered:!0,title:!0,css:{"> .title":{margin:"-2.3rem auto 1rem"}}}]}),A=function(r){var o=r.children,i=r.centered,c=r.borderColor,l=r.as,d=r.title,s=a(r,["children","centered","borderColor","as","title"]);return t(v,n({},s,{borderColor:c,centered:i,title:Boolean(d),as:l},{children:[d&&e("h3",n({className:"title"},{children:d}),void 0),o]}),void 0)},x={html:{fontSize:16,"-webkit-text-size-adjust":"100%","-webkit-tap-highlight-color":"transparent"},body:{margin:0,fontFamily:"'Press Start 2P' sans-serif",lineHeight:1.5,fontSize:16},"*, ::before, ::after":{boxSizing:"border-box",borderStyle:"solid",borderWidth:0},"button, input, optgroup, select, textarea":{"-webkit-appearance":"none",appearance:"none",verticalAlign:"middle",color:"inherit",font:"inherit",background:"transparent",padding:0,margin:0,borderRadius:0,textAlign:"inherit",textTransform:"inherit"},'button, [type="button"], [type="reset"], [type="submit"]':{cursor:"pointer"},fieldset:{margin:0,padding:0,minWidth:0},table:{borderColor:"inherit",borderCollapse:"collapse"},caption:{textAlign:"left"}},w=d(x);export{u as Button,A as Container,s as create8bitTheme,w as globalReact8BitStyles,x as globalReact8BitStylesObj};
+import{jsx as e,jsxs as t}from"react/jsx-runtime";import{rem as r}from"polished";import{createStitches as o}from"@stitches/react";var n=function(){return n=Object.assign||function(e){for(var t,r=1,o=arguments.length;r<o;r++)for(var n in t=arguments[r])Object.prototype.hasOwnProperty.call(t,n)&&(e[n]=t[n]);return e},n.apply(this,arguments)};function a(e,t){var r={};for(var o in e)Object.prototype.hasOwnProperty.call(e,o)&&t.indexOf(o)<0&&(r[o]=e[o]);if(null!=e&&"function"==typeof Object.getOwnPropertySymbols){var n=0;for(o=Object.getOwnPropertySymbols(e);n<o.length;n++)t.indexOf(o[n])<0&&Object.prototype.propertyIsEnumerable.call(e,o[n])&&(r[o[n]]=e[o[n]])}return r}var i,c={primary:"rgb(32, 156, 238)",success:"rgb(146, 204, 65)",warning:"rgb(247, 213, 29)",error:"rgb(231, 110, 85)"},l=(i=o({theme:{colors:n({},c)},media:{bp1:"(min-width: 480px)",pixelRatio:"(-webkit-min-device-pixel-ratio: 0) and (min-resolution: 0.001dpcm)"},prefix:"react8bit",utils:{marginX:function(e){return{marginLeft:e,marginRight:e}}}})).styled;i.css;var d=i.globalCss;i.keyframes,i.getCssText,i.theme;var s=i.createTheme;function f(e){return Object.keys(e)}i.config;var h={primary:{text:"#ffffff",hover:"#108de0",shade:"#006bb3"},success:{text:"#ffffff",hover:"#76c442",shade:"#4aa52e"},warning:{text:"#212529",hover:"#f2c409",shade:"#e59400"},error:{text:"#fff",hover:"#ce372b",shade:"#8c2022"}},b="#212529",g="#e7e7e7",p="#adafbc",m=l("button",{backgroundColor:"#ffffff",fontSize:"1rem",borderStyle:"solid",borderWidth:4,borderImageSlice:2,borderImageWidth:2,borderImageRepeat:"stretch",borderImageSource:'url(\'data:image/svg+xml;utf8,<?xml version="1.0" encoding="UTF-8" ?><svg version="1.1" width="5" height="5" xmlns="http://www.w3.org/2000/svg"><path d="M2 1 h1 v1 h-1 z M1 2 h1 v1 h-1 z M3 2 h1 v1 h-1 z M2 3 h1 v1 h-1 z" fill="rgb(33,37,41)" /></svg>\')',borderImageOutset:2,position:"relative",display:"inline-block",padding:"6px 8px",margin:4,textAlign:"center",verticalAlign:"middle",cursor:"url(data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAACAAAAAgCAYAAABzenr0AAAAzElEQVRYR+2X0Q6AIAhF5f8/2jYXZkwEjNSVvVUjDpcrGgT7FUkI2D9xRfQETwNIiWO85wfINfQUEyxBG2ArsLwC0jioGt5zFcwF4OYDPi/mBYKm4t0U8ATgRm3ThFoAqkhNgWkA0jJLvaOVSs7j3qMnSgXWBMiWPXe94QqMBMBc1VZIvaTu5u5pQewq0EqNZvIEMCmxAawK0DNkay9QmfFNAJUXfgGgUkLaE7j/h8fnASkxHTz0DGIBMCnBeeM7AArpUd3mz2x3C7wADglA8BcWMZhZAAAAAElFTkSuQmCC) 14 0,pointer",userSelect:"none","&[href]":{textDecoration:"none"},"&::after":{position:"absolute",top:r(-4),right:r(-4),bottom:r(-4),left:r(-4),content:"",boxShadow:"inset -".concat(r(4)," -").concat(r(4)," ").concat(p)},"&:hover":{backgroundColor:g,"&::after":{boxShadow:"inset -".concat(r(6)," -").concat(r(6)," ").concat(p)}},"&:active":{"&::after":{boxShadow:"inset ".concat(r(4)," ").concat(r(4)," ").concat(p)}},"&[disabled]":{opacity:.5,cursor:"not-allowed",backgroundColor:"#d3d3d3",color:b,"&::after":{boxShadow:"inset -".concat(r(4)," -").concat(r(4)," ").concat(p)}},"@pixelRatio":{borderImageRepeat:"space"},variants:{color:n({},function(){for(var e={},t=0,o=f(c);t<o.length;t++){var n=o[t];e[n]={"&:not([disabled])":{backgroundColor:"$".concat(n),color:h[n].text,"&::after":{boxShadow:"inset -".concat(r(4)," -").concat(r(4)," ").concat(h[n].shade)},"&:hover":{backgroundColor:h[n].hover,"&::after":{boxShadow:"inset -".concat(r(6)," -").concat(r(6)," ").concat(h[n].shade)}},"&:active":{"&::after":{boxShadow:"inset ".concat(r(4)," ").concat(r(4)," ").concat(h[n].shade)}}}}}return e}())}}),u=function(t){var r=t.children,o=t.color,i=t.href,c=a(t,["children","color","href"]);return e(m,n({type:i?void 0:"button"},c,{color:o,as:i?"a":"button",href:i},{children:r}))},v=l("section",{position:"relative",padding:"1.5rem 2rem",borderColor:"#000000",borderStyle:"solid",borderWidth:4,variants:{centered:{true:{textAlign:"center"}},title:{true:{"> .title":{display:"table",padding:"0 .5rem",margin:"-2.3rem 0 1rem",fontSize:"1rem",backgroundColor:"#ffffff"}}},borderColor:n({},function(){for(var e={},t=0,r=f(c);t<r.length;t++){var o=r[t];e[o]={borderColor:"$".concat(o)}}return e}())},compoundVariants:[{centered:!0,title:!0,css:{"> .title":{margin:"-2.3rem auto 1rem"}}}]}),A=function(r){var o=r.children,i=r.centered,c=r.borderColor,l=r.as,d=r.title,s=a(r,["children","centered","borderColor","as","title"]);return t(v,n({},s,{borderColor:c,centered:i,title:Boolean(d),as:l},{children:[d&&e("h3",n({className:"title"},{children:d})),o]}))},x={html:{fontSize:16,"-webkit-text-size-adjust":"100%","-webkit-tap-highlight-color":"transparent"},body:{margin:0,fontFamily:"'Press Start 2P' sans-serif",lineHeight:1.5,fontSize:16},"*, ::before, ::after":{boxSizing:"border-box",borderStyle:"solid",borderWidth:0},"button, input, optgroup, select, textarea":{"-webkit-appearance":"none",appearance:"none",verticalAlign:"middle",color:"inherit",font:"inherit",background:"transparent",padding:0,margin:0,borderRadius:0,textAlign:"inherit",textTransform:"inherit"},'button, [type="button"], [type="reset"], [type="submit"]':{cursor:"pointer"},fieldset:{margin:0,padding:0,minWidth:0},table:{borderColor:"inherit",borderCollapse:"collapse"},caption:{textAlign:"left"}},w=d(x);export{u as Button,A as Container,s as create8bitTheme,w as globalReact8BitStyles,x as globalReact8BitStylesObj};
 //# sourceMappingURL=index.js.map
