+++ packages/components/package.json
@@ -20,10 +20,9 @@
   },
   "peerDependencies": {
     "@stitches/react": "^1.2.5",
-    "motion": "^10.4.0",
-    "polished": "^4.1.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "polished": "^4.2.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0"
   },
   "gitHead": "de29923501826ea25154a47f539061d3efbb6114",
   "devDependencies": {
@@ -35,8 +34,7 @@
     "rollup-plugin-terser": "^7.0.2",
     "rollup-plugin-typescript2": "^0.31.1",
     "rollup-plugin-uglify": "^6.0.4",
-    "typescript": "^4.5.4",
-    "watch-run": "^1.2.5"
+    "typescript": "^4.5.4"
   },
   "dependencies": {
     "@irubingu/react-8bit-tokens": "^1.2.15-alpha.0"
