+++ packages/components/src/components/Container.tsx
@@ -1,10 +1,11 @@
 import { colorsTokens } from '@irubingu/react-8bit-tokens';
 import { CSS } from '@stitches/react';
-import type { FC } from 'react';
+import type { FC, ReactNode } from 'react';
 import { styled } from '../stitches.config';
 import typedObjectKeys from '../utils/typedObjectKeys';
 
 type React8bitUiContainerType = FC<{
+  children?: ReactNode;
   centered?: boolean;
   borderColor?: keyof typeof colorsTokens;
   bgColor?: CSS['backgroundColor'];
