+++ packages/tokens/tsconfig-build.json
@@ -1,35 +1,34 @@
 {
-    "compilerOptions": {
-      "allowSyntheticDefaultImports": true,
-      "baseUrl": "./src",
-    
-      "target": "es5",                                  /* Set the JavaScript language version for emitted 
+  "compilerOptions": {
+    "allowSyntheticDefaultImports": true,
+    "baseUrl": "./src",
+
+    "target": "es5" /* Set the JavaScript language version for emitted 
   
-      /* Modules */
-      "module": "CommonJs",                                /* Specify what module code is generated. */
-      "moduleResolution": "node",                       /* Specify how TypeScript looks up a file from                /* Allow accessing UMD globals from modules. */
-      "resolveJsonModule": true,                        /* Enable importing .json files */
-      // "noResolve": true,                                /* Disallow `import`s, `require`s or `<reference>`s from expanding the number of files TypeScript should add to a project. */
-  
-      /* JavaScript Support */
-      "allowJs": true,                                  /* Allow JavaScript files to be a part of your program. Use the `checkJS` option to get errors from these files. */
-      // "checkJs": true,                                  /* Enable error reporting in type-checked JavaScript files. */
-      // "maxNodeModuleJsDepth": 1,                        /* Specify the maximum folder depth used for checking JavaScript files from `node_modules`. Only applicable with `allowJs`. */
-  
-      /* Emit */
-      "declaration": true,                              /* Generate .d.ts files from TypeScript and JavaScript files in your project. */
-      // "declarationMap": true,                           /* Create sourcemaps for d.ts files. */
-      // "emitDeclarationOnly": true,                      /* Only output d.ts files and not JavaScript files. */
-      "sourceMap": true,                               
-      "outDir": "./build",                                   
-      "esModuleInterop": true,                            
-      "forceConsistentCasingInFileNames": true,           
-  
-      /* Type Checking */
-      "strict": true,                                      /* Enable all strict type-checking options. */
-      "skipLibCheck": true                                /* Skip type checking all .d.ts files. */
-    },
-    "include": ["./src/**/*.ts"],
-    "exclude": ["node_modules", "lib"]
-  }
-  
+      /* Modules */,
+    "module": "CommonJs" /* Specify what module code is generated. */,
+    "moduleResolution": "node" /* Specify how TypeScript looks up a file from                /* Allow accessing UMD globals from modules. */,
+    "resolveJsonModule": true /* Enable importing .json files */,
+    // "noResolve": true,                                /* Disallow `import`s, `require`s or `<reference>`s from expanding the number of files TypeScript should add to a project. */
+
+    /* JavaScript Support */
+    "allowJs": true /* Allow JavaScript files to be a part of your program. Use the `checkJS` option to get errors from these files. */,
+    // "checkJs": true,                                  /* Enable error reporting in type-checked JavaScript files. */
+    // "maxNodeModuleJsDepth": 1,                        /* Specify the maximum folder depth used for checking JavaScript files from `node_modules`. Only applicable with `allowJs`. */
+
+    /* Emit */
+    "declaration": true /* Generate .d.ts files from TypeScript and JavaScript files in your project. */,
+    // "declarationMap": true,                           /* Create sourcemaps for d.ts files. */
+    // "emitDeclarationOnly": true,                      /* Only output d.ts files and not JavaScript files. */
+    "sourceMap": true,
+    "outDir": "./build",
+    "esModuleInterop": true,
+    "forceConsistentCasingInFileNames": true,
+
+    /* Type Checking */
+    "strict": true /* Enable all strict type-checking options. */,
+    "skipLibCheck": true /* Skip type checking all .d.ts files. */
+  },
+  "include": ["./src/**/*.ts"],
+  "exclude": ["node_modules", "lib"]
+}
