+++ packages/tokens/tsconfig.json
@@ -1,28 +1,27 @@
 {
-    "compilerOptions": {
-      "allowSyntheticDefaultImports": true,
-      "resolveJsonModule": true,                        /* Enable importing .json files */
-      // "noResolve": true,                                /* Disallow `import`s, `require`s or `<reference>`s from expanding the number of files TypeScript should add to a project. */
-  
-      /* JavaScript Support */
-      "allowJs": true,                                  /* Allow JavaScript files to be a part of your program. Use the `checkJS` option to get errors from these files. */
-      // "checkJs": true,                                  /* Enable error reporting in type-checked JavaScript files. */
-      // "maxNodeModuleJsDepth": 1,                        /* Specify the maximum folder depth used for checking JavaScript files from `node_modules`. Only applicable with `allowJs`. */
-  
-      /* Emit */
-      "declaration": true,                              /* Generate .d.ts files from TypeScript and JavaScript files in your project. */
-      // "declarationMap": true,                           /* Create sourcemaps for d.ts files. */
-      // "emitDeclarationOnly": true,                      /* Only output d.ts files and not JavaScript files. */
-      "sourceMap": true,                               
-      "outDir": "./build",                                   
-      "esModuleInterop": true,                            
-      "forceConsistentCasingInFileNames": true,   
-      "emitDeclarationOnly": true,   
-  
-      /* Type Checking */
-      "strict": true,                                      /* Enable all strict type-checking options. */
-    },
-    "include": ["./build/**/*.ts"],
-    "exclude": ["node_modules"]
-  }
-    
+  "compilerOptions": {
+    "allowSyntheticDefaultImports": true,
+    "resolveJsonModule": true /* Enable importing .json files */,
+    // "noResolve": true,                                /* Disallow `import`s, `require`s or `<reference>`s from expanding the number of files TypeScript should add to a project. */
+
+    /* JavaScript Support */
+    "allowJs": true /* Allow JavaScript files to be a part of your program. Use the `checkJS` option to get errors from these files. */,
+    // "checkJs": true,                                  /* Enable error reporting in type-checked JavaScript files. */
+    // "maxNodeModuleJsDepth": 1,                        /* Specify the maximum folder depth used for checking JavaScript files from `node_modules`. Only applicable with `allowJs`. */
+
+    /* Emit */
+    "declaration": true /* Generate .d.ts files from TypeScript and JavaScript files in your project. */,
+    // "declarationMap": true,                           /* Create sourcemaps for d.ts files. */
+    // "emitDeclarationOnly": true,                      /* Only output d.ts files and not JavaScript files. */
+    "sourceMap": true,
+    "outDir": "./build",
+    "esModuleInterop": true,
+    "forceConsistentCasingInFileNames": true,
+    "emitDeclarationOnly": true,
+
+    /* Type Checking */
+    "strict": true /* Enable all strict type-checking options. */
+  },
+  "include": ["./build/**/*.ts"],
+  "exclude": ["node_modules"]
+}
